% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdata.R
\docType{data}
\name{sdata}
\alias{sdata}
\title{Simulated Irregular Longitudinal Data}
\format{
A data frame with 8631 rows and 3 variables:
\describe{
  \item{subject_id}{ID of subjects}
  \item{time}{Irregular time points.}
  \item{response}{Response values at different time points.}
}
}
\usage{
data(sdata)
}
\description{
Simulated irregular longitudinal data for 1000 patients.
This dataset contains irregularly spaced time points and responses for analysis.
}
\examples{
data(sdata)
head(sdata)
}
\keyword{datasets}
